AP-4399: Fix existing unit tests
